+++ openrouteservice-api-tests/src/test/java/heigit/ors/services/isochrones/ResultTest.java
@@ -187,14 +187,14 @@ public class ResultTest extends ServiceTest {
 			.then()
 			.body("any { it.key == 'type' }", is(true))
 			.body("any { it.key == 'features' }", is(true))
-			.body("features.size()", is(6))
+			.body("features.size()", is(5))
 			.body("features[0].type", is("Feature"))
 			.body("features[0].geometry.type", is("Polygon"))
-			.body("features[0].geometry.coordinates[0].size", is(33))
+			.body("features[0].geometry.coordinates[0].size", is(23))
 			.body("features[0].properties.containsKey('area')", is(true))
 			.body("features[1].type", is("Feature"))
 			.body("features[1].geometry.type", is("Polygon"))
-			.body("features[1].geometry.coordinates[0].size", is(52))
+			.body("features[1].geometry.coordinates[0].size", is(56))
 			.body("features[2].type", is("Feature"))
 			.body("features[2].geometry.type", is("Polygon"))
 			//.body("features[2].geometry.coordinates[0].size", is(25))
@@ -204,8 +204,38 @@ public class ResultTest extends ServiceTest {
 			//.body("features[3].geometry.coordinates[0].size", is(33))
 			.body("features[3].geometry.coordinates[0].size", is(32))
 			.body("features[4].properties.contours.size", is(2))
-			.body("features[5].properties.contours.size", is(2))
 			
 			.statusCode(200);
+	}
+
+	@Test
+	public void testSmoothingFactor() {
+		given()
+				.param("locations", getParameter("location"))
+				.param("profile", getParameter("profile"))
+				.param("range", "2000")
+				.param("range_type", "distance")
+				.param("smoothing", "10")
+				.when()
+				.get(getEndPointName())
+				.then()
+				.body("any { it.key == 'type' }", is(true))
+				.body("any { it.key == 'features' }", is(true))
+                .body("features[0].geometry.coordinates[0].size", is(52))
+				.statusCode(200);
+
+		given()
+				.param("locations", getParameter("location"))
+				.param("profile", getParameter("profile"))
+				.param("range", "2000")
+				.param("range_type", "distance")
+				.param("smoothing", "100")
+				.when()
+				.get(getEndPointName())
+				.then()
+				.body("any { it.key == 'type' }", is(true))
+				.body("any { it.key == 'features' }", is(true))
+				.body("features[0].geometry.coordinates[0].size", is(19))
+				.statusCode(200);
 	}
 }